+++ index.html
@@ -235,8 +235,7 @@
     </section>
   </template>
 
-  <script src="./js/functions.js"></script>
-  <script src="./js/data.js"></script>
+  <script type="module" src="./js/main.js"></script>
 </body>
 
 </html>
