+++ js/main.js
@@ -1,0 +1,4 @@
+import { generatePhotos } from './data';
+
+// eslint-disable-next-line no-unused-vars
+const photos = generatePhotos();
